% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_annotate.R
\name{mxr_annotate}
\alias{mxr_annotate}
\title{Annotate the significant SNPs.}
\usage{
mxr_annotate(avinput = "",
  annovar_reference_db = "~/bin/annovar/rice_msu7/",
  database_build_number = "msu7b",
  aa_matrix_file = "~/bin/annovar/rice_msu7/grantham.matrix",
  out_prefix = "", verbose = FALSE)
}
\arguments{
\item{avinput}{Make sure that the chromosome designations match those in the
reference genome annotation. (e.g. use the *.corrected_chroms file).}

\item{annovar_reference_db}{The complete path (with trailing "/") to the
annovar local database containing the reference genome.}

\item{database_build_number}{The database build number of the annovar db.}

\item{aa_matrix_file}{The grantham.matrix file or any other amino acid
substitution "relevance" file.}

\item{out_prefix}{The output prefix.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_annotate} annotates the significant SNPs using annovar.
}
\details{
Before running this function, \code{mxr_create_annovar_inputs} needs to be
run first to generate the necessary input files. Note that it is important
that the chromosome designation matches exactly the way the chromosome number
is designated in the annotation file for the reference genome. For example,
"chr01" is not the same as "Chr1", or "Chr01", moreso "1".
}

