% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_associate.R
\name{mxr_associate}
\alias{mxr_associate}
\title{Associate the genotype with the phenotype.}
\usage{
mxr_associate(tped_prefix, pheno_file, kin_file, out_prefix, cov_file = "",
  verbose = F)
}
\arguments{
\item{tped_prefix}{The path to the transposed '12' recoded plink file. The A1
allele (the minor allele -- or effect allele) is recoded as "1", while A2
allele is recoded as "2". The complete call to plink is \code{"plink2
--bfile [bed_prefix] --recode12 --output-missing-genotype 0 --transpose
--out [tped_prefix]"}.}

\item{pheno_file}{The path to the phenotype file. There are three required
columns: FID (family ID), IID (individual ID), and phenotype column. It is
recommended that the phenotype column be approximately normally
distributed, in which case, the warpedLMM software would be very useful.
IMPORTANT: The samples need to be sorted according to how they are arranged
in the genotype file. Missing phenotype values must be represented as 'NA'.}

\item{kin_file}{The path to the kinship matrix file computed using emmax-kin.
Use the Balding-Nichols algorithm. Generate the kinship matrix by a call to
emmax-kin \code{"emmax-kin-intel64 -v -d 10 [tped_prefix]"}.}

\item{out_prefix}{Path and prefix of the output files.}

\item{cov_file}{(Optional) Path to the covariate file. The first two columns
of the covariate file must be the FID and IID of the samples. The third
column must be a vector of '1'. Quantitative covariates start from the
fourth column onward.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the EMMAX run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_associate} outputs the results of the mixed linear model
association between the SNPs and the phenotype. It also outputs the REsidual
Maximun Likelihood (REML) estimates as well as the pseudoheritability
estimate. All calculations are done using EMMAX
\url{http://genome.sph.umich.edu/wiki/EMMAX}
}
\details{
Before running this function, make sure to compute EMMAX to compute the
association of each SNP, coded as minor allele additive dosage (using
--recode A in PLINK), with the phenotype.
}

