% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_clump.R
\name{mxr_clump}
\alias{mxr_clump}
\title{Post-GWAS Processing of Significant Association Peaks.}
\usage{
mxr_clump(emmax2_results, genotype_prefix, snp_field = "snp_id",
  clump_field = "p", clump_p1 = 1e-05, clump_p2 = 0.01, clump_kb = 200,
  clump_r2 = 0.8, clump_range_border_kb = 20, gene_list = "",
  out_prefix = "", verbose = FALSE)
}
\arguments{
\item{emmax2_results}{The fullpath and filename of the results of
mxr_associate.}

\item{genotype_prefix}{The fullpath and prefix of the bim/bed/fam plink
files.}

\item{snp_field}{The column header for the SNP column in the results file
from mxr_associate. (Default: "snp_id")}

\item{clump_field}{The column header for the p-value column in the results
file from mxr_associate. (Default: "p")}

\item{clump_p1}{The maximum p-value of the index variants that define each
clump. Index variants need to have values between 0 and this p-value.
(Default: 0.00001)}

\item{clump_p2}{The maximum association p-value of SNPs that belonging to the
clump formed by an index variant. (Default: 0.01)}

\item{clump_kb}{The maximum distance of a clumped SNP from its index variant.
(Default: 200)}

\item{clump_r2}{The r-squared value of the clumped SNP with its index
variant. (Default: 0.50)}

\item{clump_range_border_kb}{The number of kilobases past the region
clump-range region. (Default: 20)}

\item{gene_list}{The list of genes must have the following columns: chromosome,
start (bp), end (bp), locus_id. These fields must be tab separated and have an appropriately named column header. The chromosome must be coded to correspond to the plink data.
@param out_prefix Path and prefix of the output files.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_clump} determines the clumps of SNPs based on LD along the region
that registers a significant association signal with the phenotype. This uses
the \code{--clump} command of plink.
\url{https://www.cog-genomics.org/plink2}
}
\details{
Before running this function, make sure that the mxr_associate function has
already been called and that the results are successfully obtained without
warnings (that imply fatal errors) or errors.
}

