% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_create_haploview_inputs.R
\name{mxr_create_haploview_inputs}
\alias{mxr_create_haploview_inputs}
\title{Create the Haploview Input Files.}
\usage{
mxr_create_haploview_inputs(genotype_prefix, snps_list = "",
  out_prefix = "", verbose = FALSE)
}
\arguments{
\item{genotype_prefix}{The fullpath and prefix of the bim/bed/fam plink
files.}

\item{snps_list}{The list of clumped SNPs.}

\item{out_prefix}{Path and prefix of the output files.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_create_haploview_inputs} creates the input files necessary to run
Haploview.
}
\details{
The output of this file is required to run Haploview. It recodes in haploview
format the plink file, and then creates the batch input file that would
subsequently be fed into haploview.
}

