% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_create_haploview_plot.R
\name{mxr_create_haploview_plot}
\alias{mxr_create_haploview_plot}
\title{Generate the Haploview Plot.}
\usage{
mxr_create_haploview_plot(path_to_haploview = "", ped_file = "",
  info_file = "", max_distance = 200, memory_mb = 2048, out_prefix = "",
  verbose = FALSE)
}
\arguments{
\item{path_to_haploview}{The fullpath to the Haploview.jar.}

\item{ped_file}{The fullpath to the PED file in haploview format.}

\item{info_file}{The fullpath to the corresponding INFO file of the PED file.}

\item{max_distance}{The window size in kb of the LD block.}

\item{memory_mb}{The memory to set aside for the haploview run.
(DEFAULT:2048MB)}

\item{out_prefix}{The output prefix for the generated files.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_create_haploview_plot} generates the haploview plot using Gabriel's
algorithm.
}
\details{
For really large LD plots, it may be necessary to set the minimum memory
footprint to greater than 2GB. The default memory setting for this function
was already increased to 2GB from the default Haploview setting of 512MB.
}

