% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_create_region_file.R
\name{mxr_create_region_file}
\alias{mxr_create_region_file}
\title{Create the Region File from the Clumps.}
\usage{
mxr_create_region_file(genotype_prefix, reference_alleles = "",
  clumped_snps = "", out_prefix = "", verbose = FALSE)
}
\arguments{
\item{genotype_prefix}{The fullpath and prefix of the bim/bed/fam plink
files.}

\item{reference_alleles}{The genome-wide list of reference allele for each
SNP. It has three mandatory columns: chr, bp, Ref_allele. Chr column may
have values like "chr01"..."chr12", bp are just the base pair positions of
the SNPs, and the Ref_allele column is formatted as "Ref=<nuc>".}

\item{clumped_snps}{These are contained in the file outputted by
\code{mxr_extract_clumped_snps()}. That function needs to be run first
prior to this.}

\item{out_prefix}{Path and prefix of the output files.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_create_region_file} creates the region file that will be used to
extract the corresponding information from SNPs stored in the SNP reference
file (that contains the reference alleles collated from processing the VCF
files.)
}
\details{
The output of this file will be used in annotating the GWAS results.
}

