% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_extract_clumped_snps.R
\name{mxr_extract_clumped_snps}
\alias{mxr_extract_clumped_snps}
\title{Extract SNPs from the Clumps.}
\usage{
mxr_extract_clumped_snps(out_prefix = "", verbose = FALSE)
}
\arguments{
\item{out_prefix}{Path and prefix of the output files.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_extract_clumped_snps} creates a list of SNPs that were identified
to be part of the clumps.
}
\details{
This function needs to be run after the \code{mxr_clump} function.
}

