% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_manhattan.R, R/mxr_qq.R
\name{mxr_manhattan}
\alias{mxr_manhattan}
\title{Generate the manhattan plot.}
\usage{
mxr_manhattan(p_values = "", chr = "CHR", bp = "BP", p = "P",
  snp = "SNP", suggestive_line = -log10(1e-05),
  genomewide_line = -log10(1e-07), highlight = NULL, xlim = 10,
  ylim = 20, file_prefix = "", width = 5, height = 5, unit = "in",
  res = 300, mar = c(5, 4, 4, 0) + 0.1, mgp = c(3, 0.8, 0),
  cex.axis = 0.8, main = "", y_tickmarks = "", y_tickmark_labels = "",
  manhattan = T, qq = T)

mxr_manhattan(p_values = "", chr = "CHR", bp = "BP", p = "P",
  snp = "SNP", suggestive_line = -log10(1e-05),
  genomewide_line = -log10(1e-07), highlight = NULL, xlim = 10,
  ylim = 20, file_prefix = "", width = 5, height = 5, unit = "in",
  res = 300, mar = c(5, 4, 4, 0) + 0.1, mgp = c(3, 0.8, 0),
  cex.axis = 0.8, main = "", y_tickmarks = "", y_tickmark_labels = "",
  manhattan = T, qq = T)
}
\arguments{
\item{chr}{Header name of chromosome column.}

\item{bp}{Header for base-pair position.}

\item{p}{Header for the p-value column.}

\item{snp}{Header for the SNP ID column.}

\item{suggestive_line}{Default -log10(1e-5). Set to FALSE to disable.}

\item{highlight}{Character vector of SNPs to highlight in the manhattan plot.}

\item{xlim}{Maximum x-axis value.}

\item{ylim}{Maximum y-axis value.}

\item{file_prefix}{Fullpath plus prefix of filename of manhattan and/or qq
plot/s.}

\item{width}{Width (in inches) of the figure.}

\item{height}{Height (in inches) of the figure.}

\item{unit}{Default: "in" (inches).}

\item{res}{Resolution of the image.}

\item{mar}{Vector of margins.}

\item{mgp}{The mgp vector in plot.}

\item{cex.axis}{Same as the plot parameter.}

\item{main}{Same as the plot parameter.}

\item{y_tickmarks}{Vector of values for the y-axis.}

\item{y_tickmark_labels}{Vector of values to use as labels of the
y-tickmarks.}

\item{manhattan}{Set to TRUE to plot the manhattan plot.}

\item{qq}{Set to TRUE to plot the qq plot.}

\item{gwas_results}{The fullpath to the GWAS results. The file must have a
header line indicating the snp_id, chromosome, pvalue, and base-pair
position.}

\item{genomewideline}{Default -log10(1e-7). Set to FALSE to disable.}

\item{file}{Fullpath to the where to store the filename.}

\item{gwas_results}{The fullpath to the GWAS results. The file must have a
header line indicating the snp_id, chromosome, pvalue, and base-pair
position.}

\item{chr}{Header name of chromosome column.}

\item{bp}{Header for base-pair position.}

\item{p}{Header for the p-value column.}

\item{snp}{Header for the SNP ID column.}

\item{suggestive_line}{Default -log10(1e-5). Set to FALSE to disable.}

\item{genomewideline}{Default -log10(1e-7). Set to FALSE to disable.}

\item{highlight}{Character vector of SNPs to highlight in the manhattan plot.}

\item{ylim}{Maximum y-axis value.}

\item{width}{Width (in inches) of the figure.}

\item{height}{Height (in inches) of the figure.}

\item{unit}{Default: "in" (inches).}

\item{res}{Resolution of the image.}

\item{mar}{Vector of margins.}

\item{mgp}{The mgp vector in plot.}
}
\description{
\code{mxr_manhattan} generates the manhattan plot of the GWAS results.

\code{mxr_manhattan} generates the manhattan plot of the GWAS results.
}
\details{
Map the needed columns in the GWAS results to the manhattan function.

Map the needed columns in the GWAS results to the manhattan function.
}

