% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_summarise.R
\name{mxr_summarise}
\alias{mxr_summarise}
\title{Summarize the GWAS results into an Excel file.}
\usage{
mxr_summarise(trait = "", gwas_result = "", genotype_bim = "",
  clump_file = "", clump_ranges_file = "", clumped_genes = "",
  clumped_snps = "", reference_annotation = "", output_prefix = "",
  variant_function = "", exonic_variant_function = "", verbose = FALSE)
}
\arguments{
\item{trait}{The directory containing the trait phenotype. All GWAS results
are also stored into this directory that includes the outputs of this
function.}

\item{gwas_result}{The output of emmax2. (*.ps)}

\item{genotype_bim}{The .bim file of the genotype data.}

\item{clump_file}{The .clumped file.}

\item{clump_ranges_file}{The .clumped.ranges file.}

\item{clumped_genes}{The .clumped.genes file outputted by the plink file.}

\item{clumped_snps}{The .clumped.snps file.}

\item{reference_annotation}{The excel file containing the annotation file
that will be used to annotate the genes.}

\item{output_prefix}{Path and prefix of the output files.}

\item{variant_function}{Output of annovar.}

\item{exonic_variant_function}{Output of annovar.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_summarise} creates an Excel workbook with worksheets for the SNPs
and linked SNPs, SNPs and associated genes, annotation of SNPs, annotation of
genes, and the identified tag SNPs. A worksheet is also prepared that
contains the parameters used in the GWAS run.
}
\details{
\code{mxr_summarise} requires the results of \code{mxr::clump}.
}

