% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxr_summarise.R
\name{mxr_summarise}
\alias{mxr_summarise}
\title{Summarize the GWAS results into an Excel file.}
\usage{
mxr_summarise(trait = "", clumped_genes = "", reference_annotation = "",
  output_prefix = "", verbose = FALSE)
}
\arguments{
\item{trait}{The directory containing the trait phenotype. All GWAS results
are also stored into this directory that includes the outputs of this
function.}

\item{clumped_genes}{The clumped.genes file outputted by the plink file.}

\item{reference_annotation}{The excel file containing the annotation file
that will be used to annotate the genes.}

\item{verbose}{(Optional) Show verbose output. (DEFAULT=FALSE)}

\item{out_prefix}{Path and prefix of the output files.}
}
\value{
TRUE if the PLINK run completed successfully. FALSE, otherwise.
}
\description{
\code{mxr_summarise} creates an Excel workbook with worksheets for the SNPs
and linked SNPs, SNPs and associated genes, annotation of SNPs, annotation of
genes, and the identified tag SNPs. A worksheet is also prepared that
contains the parameters used in the GWAS run.
}
\details{
\code{mxr_summarise} requires the results of \code{mxr::clump}.
}

